requirements.txt
requests>=2.31.0
pystray>=1.10.0
Pillow>=10.0.0
